intelligent_feeding_tray/requirements.txt
opencv-python
pyserial
sounddevice
scipy
numpy
pynput
openai
requests
pydub
aiohttp